Fix Server and update dependencies
