+++ Lib/Sources/UIUnitTest/CallServer.swift
@@ -1,6 +1,7 @@
 import Foundation
 import UIKit
 import Synchronization
+import UIUnitTestAPI
 
 final class ServerAPI: Sendable {
     
@@ -142,61 +143,6 @@ func deviceId() -> Int {
     return deviceId
 }
 
-public enum Response<T: Codable> {
-    case error(error: ErrorResponse)
-    case success(data: T)
-}
 
-extension Response: Sendable where T: Sendable {
-    
-}
 
-public struct ErrorResponse: Codable, Sendable {
-    var error: String
-}
 
-public struct UIResponse<T: Codable>: Codable {
-    
-    public let response: Response<T>
-    
-    public init(response: T) {
-        self.response = .success(data: response)
-    }
-    
-    public init(error: String) {
-        self.response = .error(error: ErrorResponse(error: error))
-    }
-    
-    enum CodingKeys: CodingKey {
-        case data
-        case error
-    }
-    
-    public init(from decoder: Decoder) throws {
-        let container = try decoder.container(keyedBy: CodingKeys.self)
-        
-        if let error = try container.decodeIfPresent(ErrorResponse.self, forKey: .error) {
-            self.response = .error(error: error)
-        } else if let response = try container.decodeIfPresent(T.self, forKey: .data) {
-            self.response = .success(data: response)
-        } else {
-            fatalError("Invalid response")
-        }
-    }
-    
-    public func encode(to encoder: Encoder) throws {
-        var container = encoder.container(keyedBy: CodingKeys.self)
-        
-        switch self.response {
-        case .error(error: let error):
-            try container.encode(error, forKey: .error)
-        case .success(data: let data):
-            try container.encode(data, forKey: .data)
-        }
-    }
-    
-}
-
-extension UIResponse: Sendable where T: Sendable {
-    
-}
