+++ Lib/Sources/UIUnitTest/Element/Element+Sync.swift
@@ -1,4 +1,5 @@
 import Foundation
+import UIUnitTestAPI
 
 extension Element {
