--- Lib/Sources/UIUnitTest/Query/Query+QueryType.swift
@@ -1,85 +1,0 @@
-public extension Query {
-    enum QueryType: Int, Codable, Sendable {
-        case activityIndicators
-        case alerts
-        case browsers
-        case buttons
-        case cells
-        case checkBoxes
-        case collectionViews
-        case colorWells
-        case comboBoxes
-        case datePickers
-        case decrementArrows
-        case dialogs
-        case disclosureTriangles
-        case disclosedChildRows
-        case dockItems
-        case drawers
-        case grids
-        case groups
-        case handles
-        case helpTags
-        case icons
-        case images
-        case incrementArrows
-        case keyboards
-        case keys
-        case layoutAreas
-        case layoutItems
-        case levelIndicators
-        case links
-        case maps
-        case mattes
-        case menuBarItems
-        case menuBars
-        case menuButtons
-        case menuItems
-        case menus
-        case navigationBars
-        case otherElements
-        case outlineRows
-        case outlines
-        case pageIndicators
-        case pickerWheels
-        case pickers
-        case popUpButtons
-        case popovers
-        case progressIndicators
-        case radioButtons
-        case radioGroups
-        case ratingIndicators
-        case relevanceIndicators
-        case rulerMarkers
-        case rulers
-        case scrollBars
-        case scrollViews
-        case searchFields
-        case secureTextFields
-        case segmentedControls
-        case sheets
-        case sliders
-        case splitGroups
-        case splitters
-        case staticTexts
-        case statusBars
-        case statusItems
-        case steppers
-        case switches
-        case tabBars
-        case tabGroups
-        case tableColumns
-        case tableRows
-        case tables
-        case textFields
-        case textViews
-        case timelines
-        case toggles
-        case toolbarButtons
-        case toolbars
-        case touchBars
-        case valueIndicators
-        case webViews
-        case windows
-    }
-}
