+++ Server/Server.xcodeproj/project.xcworkspace/xcshareddata/swiftpm/Package.resolved
@@ -1,13 +1,13 @@
 {
-  "originHash" : "cec6eada3aeef1a7161603afee83d308fe58ba11bcc829d3a4ab6b1952bb2ff6",
+  "originHash" : "257633917a2c5dd97eb75e157490239d9da18ffed5d2b0e8c17aab6fe5772156",
   "pins" : [
     {
       "identity" : "flyingfox",
       "kind" : "remoteSourceControl",
       "location" : "https://github.com/swhitty/FlyingFox.git",
       "state" : {
-        "branch" : "main",
-        "revision" : "b3f2ed19bd05d1228949a5548efce06a69850143"
+        "revision" : "0482ba51ff2d6d91a7ea449c3dbdce2a78802b85",
+        "version" : "0.19.0"
       }
     },
     {
@@ -15,17 +15,8 @@
       "kind" : "remoteSourceControl",
       "location" : "https://github.com/apple/swift-argument-parser",
       "state" : {
-        "revision" : "46989693916f56d1186bd59ac15124caef896560",
-        "version" : "1.3.1"
-      }
-    },
-    {
-      "identity" : "uiunittest",
-      "kind" : "remoteSourceControl",
-      "location" : "https://github.com/BrunoMazzo/UIUnitTest.git",
-      "state" : {
-        "branch" : "accessibility-audit",
-        "revision" : "31beeddcaafcc8def22ef3404f8485f21119a4ac"
+        "revision" : "41982a3656a71c768319979febd796c6fd111d5c",
+        "version" : "1.5.0"
       }
     }
   ],
