+++ Server/ServerUITests/UITestServerTest.swift
@@ -1,9 +1,12 @@
 import XCTest
+import FlyingFox
 
 final class UITestServerTest: XCTestCase {
     
     override func record(_ issue: XCTIssue) {
-        self.server.lastIssue = issue
+        MainActor.assumeIsolated { [server] in
+            server!.lastIssue = issue
+        }
     }
     
     override func setUpWithError() throws {
@@ -30,6 +33,7 @@ final class UITestServerTest: XCTestCase {
         }
     }
     
+    @MainActor
     func deviceId() -> Int {
         let deviceName = UIDevice.current.name
