+++ UIUnitTest.xcworkspace/contents.xcworkspacedata
@@ -7,4 +7,7 @@
    <FileRef
       location = "group:Client/Client.xcodeproj">
    </FileRef>
+   <FileRef
+      location = "group:Server/Server.xcodeproj">
+   </FileRef>
 </Workspace>
